Populate ExtraNames from Names
This commit will populate the subject ExtraNames in a certificate
request from the information present in Names as long as it is not
one of the default fields. This way, we can extract non-default fields
as well as populate them in the final certificate if necessary.

It also uses the proper type for the deprecated email address
attribute in the subject.

Related to smallstep/certificates#916