Add simuator for testing TPM interactions
Opening and closing of the `TPM` has been refactored such that
there's now only one location in the code where a selection is
made between the `go-tpm` and the `go-attestation` TPMs. Only
one of them will be instantiated and used until the `TPM` is
closed (and reopened) again.

Both types support the simulator to be injected. The simulator
is currently only used for testing. The simulator doesn't cover
all code paths (yet). This can probably be improved by mocking
certain calls out, for example to return expected errors.