Add `FormatCertificateFingerprint` function
The `FormatCertificateFingerprint` allows fingerprints for SSH
certificates to be calculated. As opposed to the `FormatFingerprint`
function, the new function will always include all certificate
bytes as opposed to just the marshaled public key when calculating
the fingerprint.

The default behavior for `FormatFingerprint` now behaves like other
systems, such as `ssh-add -l`.