Use renamed module fork `github.com/smallstep/go-attestation`
Downstream users of packages that depend on `crypto` will hit
issues when using a replace in this package. We've created a
branch called `surrogate` in `github.com/smallstep/go-attestation`,
which includes all our latest changes and which has its module
name renamed, allowing for a direct import instead of a replace.