Add support for (default) Windows store and location configuration
Certificates can now be stored in specific locations (user vs. machine)
and stores (e.g. "My" and "CA"). The defaults for a specific TPMKMS
instance can be provided at instantiation time. The certs for specific
keys can be stored in different locations and stores by providing the
option values as part of the certificate operations.