Return a sentinel error when no storage is configured when required
This commit changes the behavior of the blackhole storage. It now
returns a sentinel error for operations that require a storage to
be successful. The sentinel error is handled in the `tpm` package
wherever a storage operation takes place.

The goal of the blackhole storage is to be able to use an instance
of `tpm.TPM` for operations that don't require an (metadata) storage.
In some usages it could result in an error, because it's not clear
enough that some operations require an actual storage implementation
for the operations to succeed. This should now be more explicit
through the (sentinel) errors.