Accept nonRepudiation when unmarshalling keyUsage
ITU-T X.509 (10/2019) section 9.2.2.3 [1] defines 'contentCommitment' as
the current name for what had previously been called 'nonRepudiation',
and deprecates the old name:

> It is not incorrect to refer to this keyUsage bit using the identifier
> nonRepudiation. However, the use of this identifier has been
> deprecated.

The old name is still widely used, though — for example, OpenSSL still
supports nonRepudiation exclusively. It is therefore not surprising that
users attempt to specify that and just get an error message in return.
See for example smallstep/certificates#2348.

Modify the JSON unmarshalling to accept the old name, and transparently
convert it to the new one. Keep marshalling the same, since we want to
be conservative in what we send, but liberal in what we accept.

[1]: https://www.itu.int/rec/T-REC-X.509-201910-I/en

Related: smallstep/certificates#2348
Signed-off-by: Clemens Lang <neverpanic@gmail.com>